+++ CHANGELOG.md
@@ -7,13 +7,13 @@
 * *BREAKING* Implement central ordering settings for receiving search configuration. Refs UIOR-1232.
 * *BREAKING* Apply changes in the fund schema for the locations field. Refs UIOR-1251.
 * *BREAKING* Align the `finance.*` interfaces versions in accordance with the changes in the descriptor. Refs UIOR-1260.
+* Settings - implement Routing address configuration. Refs UIOR-1261.
 
 ## [6.0.3](https://github.com/folio-org/ui-orders/tree/v6.0.3) (2024-04-18)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v6.0.2...v6.0.3)
 
 * Add additional order filters to support reporting requirements. Refs UIOR-1256.
 * Eliminate shadowing when passing an undefined prop to the receipt status field. Refs UIOR-1257.
-* Settings - implement Routing address configuration. Refs UIOR-1261.
 
 ## [6.0.2](https://github.com/folio-org/ui-orders/tree/v6.0.2) (2024-04-01)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v6.0.1...v6.0.2)